+++ src/tko/play/play_actions.py
@@ -62,13 +62,13 @@ class PlayActions:
                 opener.load_folders_and_open()
             else:
                 self.fman.add_input(
-                    Floating("v>")
+                    Floating(self.settings, "v>")
                     .put_text("\nO arquivo de código não foi encontrado.\n")
                     .error()
                 )
         else:
             self.fman.add_input(
-                Floating("v>")
+                Floating(self.settings, "v>")
                 .put_text("\nVocê só pode abrir o código")
                 .put_text("de tarefas baixadas.\n")
                 .error()
@@ -84,14 +84,14 @@ class PlayActions:
                 except Exception as _:
                     pass
             self.fman.add_input(
-                Floating("v>")
+                Floating(self.settings, "v>")
                 .set_header(" Abrindo link ")
                 .put_text("\n " + task.link + " \n")
                 .warning()
             )
         elif isinstance(obj, Quest):
             self.fman.add_input(
-                Floating("v>")
+                Floating(self.settings, "v>")
                 .put_text("\nEssa é uma missão.")
                 .put_text("\nVocê só pode abrir o link")
                 .put_text("de tarefas.\n")
@@ -99,7 +99,7 @@ class PlayActions:
             )
         else:
             self.fman.add_input(
-                Floating("v>")
+                Floating(self.settings, "v>")
                 .put_text("\nEsse é um grupo.")
                 .put_text("\nVocê só pode abrir o link")
                 .put_text("de tarefas.\n")
@@ -122,7 +122,7 @@ class PlayActions:
         
         if isinstance(obj, Task):
             self.fman.add_input(
-                FloatingGrade(obj, "").set_exit_fn(
+                FloatingGrade(obj, self.settings, "").set_exit_fn(
                     lambda: self.register_action(obj)
                 )
             )
@@ -134,7 +134,7 @@ class PlayActions:
         
         if isinstance(obj, Quest):
             self.fman.add_input(
-                Floating("v>")
+                Floating(self.settings, "v>")
                 .put_text("\nEssa é uma missão.")
                 .put_text("\nVocê só pode baixar tarefas.\n")
                 .error()
@@ -142,7 +142,7 @@ class PlayActions:
             return
         if isinstance(obj, Cluster):
             self.fman.add_input(
-                Floating("v>")
+                Floating(self.settings, "v>")
                 .put_text("\nEsse é um grupo.")
                 .put_text("\nVocê só pode baixar tarefas.\n")
                 .error()
@@ -154,12 +154,12 @@ class PlayActions:
     def __down_remote_task(self, task: Task) -> None:
         if task.link_type != Task.Types.REMOTE_FILE and task.link_type != Task.Types.IMPORT_FILE:
             self.fman.add_input(
-                Floating("v>").put_text("\nEssa não é uma tarefa de baixável.\n").error()
+                Floating(self.settings, "v>").put_text("\nEssa não é uma tarefa de baixável.\n").error()
             )
             return
         lang = self.rep.get_lang() 
         down_frame = (
-            Floating("v>").warning().set_text_ljust().set_header(" Baixando tarefa ")
+            Floating(self.settings, "v>").warning().set_text_ljust().set_header(" Baixando tarefa ")
         )
         # down_frame.put_text(f"\ntko down {self.rep.alias} {task.key} -l {lang}\n")
         self.fman.add_input(down_frame)
@@ -217,7 +217,7 @@ class PlayActions:
         
         if not run.wdir.has_solver():
             DownProblem.create_default_draft(task_dir, self.rep.get_lang())
-            msg = Floating("v>").warning()
+            msg = Floating(self.settings, "v>").warning()
             msg.put_text("\nNenhum arquivo de código na linguagem {} encontrado.".format(self.rep.get_lang()))
             msg.put_text("\nUm arquivo de rascunho foi criado em {}.".format(task_dir))
             self.fman.add_input(msg)